Language.Javascript.JSC.Classes

Description

These classes are used to make various JavaScriptCore types out of whatever we have. Functions in jsc take these as inputs. This alows implicit casting and eager evaluation.

Type classes to convert Haskell data to JavaScript

class MakeValueRef a where

Anything that can be used to make a JavaScript value reference

Methods

makeValueRef :: a -> JSC JSValueRef

class MakeStringRef a where

Anything that can be used to make a JavaScript string reference

Methods

makeStringRef :: a -> JSStringRef

class MakeArgRefs this where

Anything that can be used to make a list of JavaScript value references for use as function arguments

Methods

makeArgRefs :: this -> JSC [JSValueRef]

class MakeObjectRef this where

Anything that can be used to make a JavaScript object reference

Methods

makeObjectRef :: this -> JSC JSObjectRef

class MakePropRef this where

Anything that can be used to make a JavaScript property reference

Methods

makePropRef :: this -> JSC JSPropRef