Language.Javascript.JSC.Value

Description

Deals with JavaScript values. These can be

null

undefined

true | false

a double precision floating point number

a string

an object

JavaScript value references

type JSValueRef = Ptr OpaqueJSValue

class MakeValueRef a where

Anything that can be used to make a JavaScript value reference

Methods

makeValueRef :: a -> JSC JSValueRef

Haskell types for JavaScript values

data JSNull

Constructors

JSNull
Type that represents a value that can only be null. Haskell of course has no null so we are adding this type.

type JSUndefined

Arguments

= ()
A type that can only be undefined in JavaScript. Using () because functions in JavaScript that have no return, impicitly return undefined.

type JSBool

Arguments

= Bool
JavaScript boolean values map the Bool haskell type.

type JSNumber

Arguments

= Double
A number in JavaScript maps nicely to Double.

type JSString

Arguments

= Text
JavaScript strings can be represented with the Haskell Text type.

data JSValue

An algebraic data type that can represent a JavaScript value. Any JavaScriptCore JSValueRef can be converted into this type.

Constructors

ValNull
null

ValUndefined
undefined

ValBool JSBool
true or false

ValNumber JSNumber
a number

ValString JSString
a string

ValObject JSObjectRef
an object

Converting JavaScript values

valToBool :: MakeValueRef val => val -> JSC JSBool

Given a JavaScript value get its boolean value. All values in JavaScript convert to bool.

>>> runjs $ valToBool JSNull
False
>>> runjs $ valToBool ()
False
>>> runjs $ valToBool True
True
>>> runjs $ valToBool False
False
>>> runjs $ valToBool (1.0 :: Double)
True
>>> runjs $ valToBool (0.0 :: Double)
False
>>> runjs $ valToBool ""
False
>>> runjs $ valToBool "1"
True

valToNumber :: MakeValueRef val => val -> JSC JSNumber

Given a JavaScript value get its numeric value. May throw JSException.

>>> runjs $ valToNumber JSNull
0.0
>>> runjs $ valToNumber ()
NaN
>>> runjs $ valToNumber True
1.0
>>> runjs $ valToNumber False
0.0
>>> runjs $ valToNumber (1.0 :: Double)
0.0
>>> runjs $ valToNumber (0.0 :: Double)
0.0
>>> runjs $ valToNumber ""
0.0
>>> runjs $ valToNumber "1"
1.0

valToStr :: MakeValueRef val => val -> JSC JSStringRef

Given a JavaScript value get its string value (as a JavaScript string). May throw JSException.

>>> runjs $ valToStr JSNull >>= strToText
"null"
>>> runjs $ valToStr () >>= strToText
"undefined"
>>> runjs $ valToStr True >>= strToText
"true"
>>> runjs $ valToStr False >>= strToText
"false"
>>> runjs $ valToStr (1.0 :: Double) >>= strToText
"1"
>>> runjs $ valToStr (0.0 :: Double) >>= strToText
"0"
>>> runjs $ valToStr "" >>= strToText
""
>>> runjs $ valToStr "1" >>= strToText
"1"

valToObject :: MakeValueRef val => val -> JSC JSObjectRef

Given a JavaScript value get its object value. May throw JSException.

>>> runjs $ (valToObject JSNull >>= valToText) `catch` \ (JSException e) -> valToText e
"TypeError: 'null' is not an object"
>>> runjs $ (valToObject () >>= valToText) `catch` \ (JSException e) -> valToText e
"TypeError: 'undefined' is not an object"
>>> runjs $ valToObject True
"true"
>>> runjs $ valToObject False
"false"
>>> runjs $ valToObject (1.0 :: Double)
"1"
>>> runjs $ valToObject (0.0 :: Double)
"0"
>>> runjs $ valToObject ""
""
>>> runjs $ valToObject "1"
"1"

valToText :: MakeValueRef val => val -> JSC Text

Given a JavaScript value get its string value (as a Haskell Text). May throw JSException.

>>> runjs $ valToText JSNull
"null"
>>> runjs $ valToText ()
"undefined"
>>> runjs $ valToText True
"true"
>>> runjs $ valToText False
"false"
>>> runjs $ valToText (1.0 :: Double)
"1"
>>> runjs $ valToText (0.0 :: Double)
"0"
>>> runjs $ valToText ""
""
>>> runjs $ valToText "1"
"1"

Make JavaScript values from Haskell ones

val

Arguments

:: MakeValueRef value

=> value
value to convert to a JavaScript value

-> JSC JSValueRef

Convert to a JavaScript value (just an alias for makeValueRef)

valMakeNull :: JSC JSValueRef

Make a null JavaScript value

valMakeUndefined :: JSC JSValueRef

Make an undefined JavaScript value

valMakeBool :: JSBool -> JSC JSValueRef

Make a JavaScript boolean value

valMakeNumber :: JSNumber -> JSC JSValueRef

Make a JavaScript number

valMakeString :: Text -> JSC JSValueRef

Make a JavaScript string

Conver to and from JSValue

deRefVal :: MakeValueRef val => val -> JSC JSValue

Derefernce a value reference.

>>> runjs $ deRefVal JSNull
ValNull
>>> runjs $ deRefVal ()
ValUndefined
>>> runjs $ deRefVal True
ValBool True
>>> runjs $ deRefVal False
ValBool False
>>> runjs $ deRefVal (1.0 :: Double)
ValNumber 1.0
>>> runjs $ deRefVal (0.0 :: Double)
ValNumber 0.0
>>> runjs $ deRefVal ""
ValString ""
>>> runjs $ deRefVal "1"
ValString "1"
>>> runjs $ valToObject True >>= deRefVal
ValObject 0x...

valMakeRef :: JSValue -> JSC JSValueRef

Make a JavaScript value out of a JSValue ADT.

>>> runjs $ valMakeRef ValNull
"null"
>>> runjs $ valMakeRef ValUndefined
"undefined"
>>> runjs $ valMakeRef (ValBool True)
"true"
>>> runjs $ valMakeRef (ValNumber 1)
"1"
>>> runjs $ valMakeRef (ValString $ pack "Hello")
"Hello"